Abstracts labeling fmethods used by Issue and Repo into Lableable concern.
