+++ app/models/issue.rb
@@ -1,24 +1,12 @@
 class Issue < ActiveRecord::Base
-  self.inheritance_column = nil
-
-  # [TODO] CJL // 2015-03-12
-  # Worth moving the label methods into a concern?
-  # Is Lableable something multiple classes want?
-  # => Look @ issue.rb && github_utils.rb
-
-  include GithubUtils::Parser
-
-  ISSUE_CATEGORIES = ['category', 'priority', 'status', 'size', 'team', 'type']
+  include Labelable
 
   has_and_belongs_to_many :labels
   belongs_to :repo
 
-  def update_label_info(repo_id, labels)
-    if labels.present?
-      self.add_labels(repo_id, labels)
-      self.add_custom_attributes(labels)
-    end
-  end
+  self.inheritance_column = nil
+
+  ISSUE_CATEGORIES = ['category', 'priority', 'status', 'size', 'team', 'type']
 
   def add_custom_attributes(labels)
     labels.each do |label|
@@ -27,13 +15,8 @@ class Issue < ActiveRecord::Base
     end
   end
 
-  def clear_labels
-    self.labels.clear
-  end
-
   def clear_custom_attributes(issue)
     ISSUE_CATEGORIES.each do |custom_category|
-      custom_category = 'issue_type' if custom_category == 'type'
       custom_categories = {}
       custom_categories[custom_category.to_sym] = nil
       issue.update_attributes(custom_categories)
