+++ app/models/repo.rb
@@ -1,5 +1,6 @@
  class Repo < ActiveRecord::Base
   include GithubUtils::Parser
+  include Labelable
 
   has_and_belongs_to_many :boards
   has_many :issues
